+++ src/appleseed/renderer/kernel/lighting/lighttracing/lighttracingsamplegenerator.cpp
@@ -561,14 +561,14 @@ namespace
         {
             // Sample the light sources.
             sampling_context.split_in_place(4, 1);
-            const Vector4d s = sampling_context.next2<Vector4d>();
+            const Vector4f s = sampling_context.next2<Vector4f>();
             LightSample light_sample;
             m_light_sampler.sample(
                 ShadingRay::Time::create_with_normalized_time(
-                    s[0],
+                    static_cast<double>(s[0]),
                     m_shutter_open_time,
                     m_shutter_close_time),
-                Vector3d(s[1], s[2], s[3]),
+                Vector3f(s[1], s[2], s[3]),
                 light_sample);
 
             return
