+++ src/appleseed/renderer/kernel/lighting/sppm/sppmphotontracer.cpp
@@ -312,14 +312,14 @@ namespace
             const ShadingContext&   shading_context,
             SamplingContext&        sampling_context)
         {
-            const Vector4d s = sampling_context.next2<Vector4d>();
+            const Vector4f s = sampling_context.next2<Vector4f>();
             LightSample light_sample;
             m_light_sampler.sample(
                 ShadingRay::Time::create_with_normalized_time(
-                    s[0],
+                    static_cast<double>(s[0]),
                     m_shutter_open_time,
                     m_shutter_close_time),
-                Vector3d(s[1], s[2], s[3]),
+                Vector3f(s[1], s[2], s[3]),
                 light_sample);
 
             if (light_sample.m_triangle)
